Add smoketest for lgt92 Add default LED declaration so example compiles Remove useless hardware init and comments Move all machine_stm32l0x2 code back to machine_stm32l0 as it'll probably work for stm32l0x1 Move Flash-method,linkerscript... from chip target to board target json
